refactor(gh-deployment/dagger/main.go): change deployment ID type from int to int64
The type of the deployment ID variable `dplId` and the return type of
the `CreateGithubDeployment` function are changed from `int` to `int64`.

+++ gh-deployment/dagger/main.go
@@ -45,8 +45,8 @@ func (ghd *GhDeployment) CreateGithubDeployment(
 	ctx context.Context,
 	// Github token for making api requests
 	token string,
-) (int, error) {
-	var dplId int
+) (int64, error) {
+	var dplId int64
 
 	owner, repo, err := parseOwnerRepo(ghd.Repository)
 	if err != nil {
@@ -96,7 +96,7 @@ func (ghd *GhDeployment) CreateGithubDeployment(
 	if err != nil {
 		return dplId, fmt.Errorf("error in creating deployment %s", err)
 	}
-	return int(dpl.GetID()), nil
+	return dpl.GetID(), nil
 }
 
 // WithRepository sets the GitHub repository name
